Create separate tests for ComWrappers feature
