+++ src/tests/nativeaot/SmokeTests/ComWrappers/CMakeLists.txt
@@ -1,0 +1,7 @@
+project (ComWrappersNative)
+include_directories(${INC_PLATFORM_DIR})
+
+add_library (ComWrappersNative SHARED ComWrappersNative.cpp)
+
+# add the install targets
+install (TARGETS ComWrappersNative DESTINATION bin)
