+++ src/tests/nativeaot/SmokeTests/PInvoke/PInvoke.cs
@@ -209,9 +209,6 @@ namespace PInvokeTests
         [DllImport("PInvokeNative", CallingConvention = CallingConvention.StdCall)]
         static extern bool IsNULL(SequentialStruct[] foo);
 
-        [DllImport("PInvokeNative", CallingConvention = CallingConvention.StdCall)]
-        static extern bool IsNULL(IComInterface foo);
-
         [StructLayout(LayoutKind.Sequential, CharSet= CharSet.Ansi, Pack = 4)]
         public unsafe struct InlineArrayStruct
         {
@@ -296,13 +293,6 @@ namespace PInvokeTests
             TestAsAny();
             TestMarshalStructAPIs();
             TestForwardDelegateWithUnmanagedCallersOnly();
-
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                TestComInteropNullPointers();
-                TestComInteropRegistrationRequired();
-                TestComInteropReleaseProcess();
-            }
 
             return 100;
         }
@@ -1009,68 +999,11 @@ namespace PInvokeTests
             GC.Collect();
         }
 
-        [DllImport("PInvokeNative", CallingConvention = CallingConvention.StdCall)]
-        static extern int CaptureComPointer(IComInterface foo);
-
-        [DllImport("PInvokeNative", CallingConvention = CallingConvention.StdCall)]
-        static extern void ReleaseComPointer();
-
         public static unsafe void TestForwardDelegateWithUnmanagedCallersOnly()
         {
             Console.WriteLine("Testing Forward Delegate with UnmanagedCallersOnly");
             Action a = Marshal.GetDelegateForFunctionPointer<Action>((IntPtr)(void*)(delegate* unmanaged<void>)&UnmanagedCallback);
             a();
-        }
-
-        public static void TestComInteropNullPointers()
-        {
-            Console.WriteLine("Testing Marshal APIs for COM interfaces");
-            IComInterface comPointer = null;
-            var result = IsNULL(comPointer);
-            ThrowIfNotEquals(true, IsNULL(comPointer), "COM interface marshalling null check failed");
-        }
-
-        public static void TestComInteropRegistrationRequired()
-        {
-            Console.WriteLine("Testing COM Interop registration process");
-            ComObject target = new ComObject();
-            try
-            {
-                CaptureComPointer(target);
-                throw new Exception("Cannot work without ComWrappers.RegisterForMarshalling called");
-            }
-            catch (InvalidOperationException)
-            {
-            }
-        }
-
-        public static void TestComInteropReleaseProcess()
-        {
-            Console.WriteLine("Testing CCW release process");
-            ComWrappers wrapper = new SimpleComWrapper();
-            ComWrappers.RegisterForMarshalling(wrapper);
-            WeakReference comPointerHolder = CreateComReference();
-
-            GC.Collect();
-            ThrowIfNotEquals(true, comPointerHolder.IsAlive, ".NET object should be alive");
-
-            ReleaseComPointer();
-
-            GC.Collect();
-            ThrowIfNotEquals(false, comPointerHolder.IsAlive, ".NET object should be disposed by then");
-        }
-
-        [MethodImpl(MethodImplOptions.NoInlining)]
-        private static WeakReference CreateComReference()
-        {
-            ComObject target = new ComObject();
-            WeakReference comPointerHolder = new WeakReference(target);
-
-            int result = CaptureComPointer(target);
-            ThrowIfNotEquals(0, result, "Seems to be COM marshalling behave stragerly.");
-            ThrowIfNotEquals(11, target.TestResult, "Call to method should work");
-
-            return comPointerHolder;
         }
     }
 
@@ -1099,69 +1032,6 @@ namespace PInvokeTests
             return ReleaseMemory(handle);
         }
     } //end of SafeMemoryHandle class
-
-    [ComImport]
-    [ComVisible(true)]
-    [Guid("111e91ef-1887-4afd-81e3-70cf08e715d8")]
-    [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-    public interface IComInterface
-    {
-        int DoWork(int param);
-    }
-
-    public class ComObject : IComInterface
-    {
-        public int TestResult;
-        public int DoWork(int param)
-        {
-            this.TestResult += param;
-            return 0;
-        }
-    }
-
-    internal unsafe class SimpleComWrapper : ComWrappers
-    {
-        static ComInterfaceEntry* wrapperEntry;
-
-        static SimpleComWrapper()
-        {
-            IntPtr* vtbl = (IntPtr*)RuntimeHelpers.AllocateTypeAssociatedMemory(typeof(IComInterface), 4 * sizeof(IntPtr));
-            GetIUnknownImpl(out vtbl[0], out vtbl[1], out vtbl[2]);
-            vtbl[3] = (IntPtr)(delegate* unmanaged<IntPtr, int, int>)&IComInterfaceProxy.DoWork;
-
-            var comInterfaceEntryMemory = RuntimeHelpers.AllocateTypeAssociatedMemory(typeof(IComInterface), sizeof(ComInterfaceEntry));
-            wrapperEntry = (ComInterfaceEntry*)comInterfaceEntryMemory;
-            wrapperEntry->IID = new Guid("111e91ef-1887-4afd-81e3-70cf08e715d8");
-            wrapperEntry->Vtable = (IntPtr)vtbl;
-        }
-
-        protected override unsafe ComInterfaceEntry* ComputeVtables(object obj, CreateComInterfaceFlags flags, out int count)
-        {
-            if (obj is not IComInterface)
-                throw new Exception();
-            count = 1;
-            return wrapperEntry;
-        }
-
-        protected override object CreateObject(IntPtr externalComObject, CreateObjectFlags flags)
-        {
-            return null;
-        }
-
-        protected override void ReleaseObjects(System.Collections.IEnumerable objects)
-        {
-        }
-    }
-
-    internal unsafe class IComInterfaceProxy
-    {
-        [UnmanagedCallersOnly]
-        public static int DoWork(IntPtr thisPtr, int param)
-        {
-            var inst = ComWrappers.ComInterfaceDispatch.GetInstance<IComInterface>((ComWrappers.ComInterfaceDispatch*)thisPtr);
-            return inst.DoWork(param);
-        }
-    }
 
     public static class LowLevelExtensions
     {
